Release 1.1.4
### Introduction 
This release includes the following changes to the API, and internal checks. This may change overtime because of bugs that are encountered in several versions. Your feedback is highly needed, and necessary. If you encountered any bugs or issues, please dont hesitate to contact us on discord or create an issue on github.

### Changes
- Added new icon for poggit release.
- Fix miscalculation of thresholds.
- Fixed Timer (A, B)
- Lessen Speed (A, B) debug spam.
- Added ClickTP (A) and Fixed Speed (A, B) not detecting.
- Fix All Scaffold Modules
- Fix issues on #19, #22 
- Fix crash in several modules.
- Fix checks for `$player->hasNoClientPredictions()`
- Added ImpossiblePitch (A)
- Added MessageSpoof (A)
- Add `{error_code}` variable in the messages.
- Fix crash on listener.

### Todo
- Discord Webhook API Fix
- Discord Embeds
- Custom messages (configurable)