Release v1.1.2
### Introduction
This update will solve various problems that have shown in previous poggit versions, which are now out of date. This includes a hotfix for the following modules and issues/bugs. There were no modules added to this version. **If you encounter any bugs or issues, please report them on our github.** It is advised that you update Zuri to ensure the most recent patch fix.

### Changes
- Hack-Fix for #32
- Fix array bug in `Regen (B)`
- Remove some of duplicates in configuration.
- Fix #26: Fastbow (A) array bug..
- Fix fly module is not working...
- Check chunk if it is loaded to prevent false-kick of the player..
- Fix crash in Webhook (#32)